refactor(ParametersRoutes): move file to its own directory
