+++ src/components/router/Router.tsx
@@ -6,7 +6,7 @@ import ValidateAuthentication from '~components/auth/Authentication'
 import Home from '~/pages/home/Home'
 import AuthenticatedLayout from '~/components/layouts/AuthenticatedLayout/AuthenticatedLayout'
 import { analysisRoutes } from './AnalysisRoutes/AnalysisRoutes'
-import { parametersRoutes } from './parametersRoutes'
+import { parametersRoutes } from './ParametersRoutes/ParametersRoutes'
 import { reportsRoutes } from './reportsRoutes'
 import { transactionsRoute } from './transactionsRoutes'
 import { usersRoutes } from './usersRoutes'
